// types/quiz.ts
export type CraftCategory =
  // 織物 (38品目)
  | "二風谷アットゥシ"
  | "置賜紬"
  | "羽越しな布"
  | "奥会津昭和からむし織"
  | "結城紬"
  | "伊勢崎絣"
  | "桐生織"
  | "秩父銘仙"
  | "村山大島紬"
  | "本場黄八丈"
  | "多摩織"
  | "塩沢紬"
  | "小千谷縮"
  | "小千谷紬"
  | "本塩沢"
  | "十日町絣"
  | "十日町明石ちぢみ"
  | "信州紬"
  | "牛首紬"
  | "近江上布"
  | "西陣織"
  | "弓浜絣"
  | "阿波正藍しじら織"
  | "博多織"
  | "久留米絣"
  | "本場大島紬"
  | "久米島紬"
  | "宮古上布"
  | "読谷山花織"
  | "読谷山ミンサー"
  | "琉球絣"
  | "首里織"
  | "与那国織"
  | "喜如嘉の芭蕉布"
  | "知花花織"
  | "南風原花織"
  | "八重山ミンサー"
  | "八重山上布"
  // 染色品 (14品目)
  | "東京染小紋"
  | "東京手描友禅"
  | "東京無地染"
  | "東京本染注染"
  | "加賀友禅"
  | "有松・鳴海絞"
  | "名古屋友禅"
  | "名古屋黒紋付染"
  | "京鹿の子絞"
  | "京友禅"
  | "京小紋"
  | "京黒紋付染"
  | "浪華本染め"
  | "琉球びんがた"
  // その他の繊維製品 (5品目)
  | "行田足袋"
  | "加賀繍"
  | "伊賀くみひも"
  | "京繡"
  | "京くみひも"
  // 陶磁器 (33品目)
  | "大堀相馬焼"
  | "会津本郷焼"
  | "笠間焼"
  | "益子焼"
  | "佐渡無名異焼"
  | "九谷焼"
  | "美濃焼"
  | "常滑焼"
  | "赤津焼"
  | "瀬戸染付焼"
  | "三州鬼瓦工芸品"
  | "四日市萬古焼"
  | "伊賀焼"
  | "越前焼"
  | "信楽焼"
  | "京焼・清水焼"
  | "丹波立杭焼"
  | "出石焼"
  | "石見焼"
  | "備前焼"
  | "萩焼"
  | "大谷焼"
  | "砥部焼"
  | "小石原焼"
  | "上野焼"
  | "伊万里・有田焼"
  | "唐津焼"
  | "三川内焼"
  | "波佐見焼"
  | "小代焼"
  | "天草陶磁器"
  | "薩摩焼"
  | "壺屋焼"
  // 漆器 (23品目)
  | "津軽塗"
  | "秀衡塗"
  | "浄法寺塗"
  | "鳴子漆器"
  | "川連漆器"
  | "会津塗"
  | "鎌倉彫"
  | "小田原漆器"
  | "村上木彫堆朱"
  | "新潟漆器"
  | "木曽漆器"
  | "高岡漆器"
  | "輪島塗"
  | "山中漆器"
  | "金沢漆器"
  | "飛騨春慶"
  | "越前漆器"
  | "若狭塗"
  | "京漆器"
  | "紀州漆器"
  | "大内塗"
  | "香川漆器"
  | "琉球漆器"
  // 木工品・竹工品 (33品目)
  | "二風谷イタ"
  | "岩谷堂箪笥"
  | "仙台箪笥"
  | "樺細工"
  | "大館曲げわっぱ"
  | "秋田杉桶樽"
  | "奥会津編み組細工"
  | "春日部桐箪笥"
  | "江戸和竿"
  | "江戸指物"
  | "箱根寄木細工"
  | "加茂桐箪笥"
  | "松本家具"
  | "南木曽ろくろ細工"
  | "駿河竹千筋細工"
  | "井波彫刻"
  | "一位一刀彫"
  | "岐阜和傘"
  | "名古屋桐箪笥"
  | "越前箪笥"
  | "京指物"
  | "大阪欄間"
  | "大阪唐木指物"
  | "大阪泉州桐箪笥"
  | "大阪金剛簾"
  | "豊岡杞柳細工"
  | "紀州箪笥"
  | "紀州へら竿"
  | "勝山竹細工"
  | "宮島細工"
  | "別府竹細工"
  | "都城大弓"
  | "高山茶筌"
  // 金工品 (16品目)
  | "南部鉄器"
  | "山形鋳物"
  | "千葉工匠具"
  | "東京銀器"
  | "東京アンチモニー工芸品"
  | "燕鎚起銅器"
  | "越後与板打刃物"
  | "越後三条打刃物"
  | "信州打刃物"
  | "高岡銅器"
  | "越前打刃物"
  | "堺打刃物"
  | "大阪浪華錫器"
  | "播州三木打刃物"
  | "土佐打刃物"
  | "肥後象がん"
  // 仏壇・仏具 (17品目)
  | "山形仏壇"
  | "長岡仏壇"
  | "三条仏壇"
  | "新潟・白根仏壇"
  | "飯山仏壇"
  | "金沢仏壇"
  | "七尾仏壇"
  | "三河仏壇"
  | "名古屋仏壇"
  | "尾張仏具"
  | "彦根仏壇"
  | "京仏壇"
  | "京仏具"
  | "大阪仏壇"
  | "広島仏壇"
  | "八女福島仏壇"
  | "川辺仏壇"
  // 和紙 (9品目)
  | "内山紙"
  | "越中和紙"
  | "美濃和紙"
  | "越前和紙"
  | "因州和紙"
  | "石州和紙"
  | "阿波和紙"
  | "大洲和紙"
  | "土佐和紙"
  // 文具 (10品目)
  | "雄勝硯"
  | "豊橋筆"
  | "鈴鹿墨"
  | "播州そろばん"
  | "奈良筆"
  | "奈良墨"
  | "雲州そろばん"
  | "熊野筆"
  | "川尻筆"
  | "赤間硯"
  // 石工品 (4品目)
  | "真壁石燈籠"
  | "岡崎石工品"
  | "京石工芸品"
  | "出雲石燈ろう"
  // 貴石細工 (2品目)
  | "甲州水晶貴石細工"
  | "若狭めのう細工"
  // 人形・こけし (10品目)
  | "宮城伝統こけし"
  | "江戸木目込人形"
  | "岩槻人形"
  | "江戸節句人形"
  | "江戸押絵"
  | "駿河雛具"
  | "駿河雛人形"
  | "名古屋節句飾"
  | "京人形"
  | "博多人形"
  // その他の工芸品 (26品目)
  | "天童将棋駒"
  | "房州うちわ"
  | "江戸からかみ"
  | "江戸切子"
  | "江戸木版画"
  | "江戸硝子"
  | "江戸べっ甲"
  | "東京三味線"
  | "東京琴"
  | "江戸表具"
  | "甲州印伝"
  | "甲州手彫印章"
  | "岐阜提灯"
  | "尾張七宝"
  | "越中福岡の菅笠"
  | "京扇子"
  | "京うちわ"
  | "京表具"
  | "いずみガラス"
  | "播州毛鉤"
  | "福山琴"
  | "丸亀うちわ"
  | "八女提灯"
  | "長崎べっ甲"
  | "山鹿灯籠"
  | "三線"
  // 工芸材料・工芸用具 (3品目)
  | "庄川挽物木地"
  | "金沢箔"
  | "伊勢形紙"

export type KnowledgeCategory = "history" | "geography" | "materials" | "usage" | "design" | "trivia"

export type QuestionFormat = "multiple_choice" | "true_false"

export type Difficulty = "easy" | "medium" | "hard"

export type GenreKey =
  | "all"
  | "textiles"
  | "dyedGoods"
  | "otherTextiles"
  | "pottery"
  | "lacquerware"
  | "woodBamboo"
  | "metalwork"
  | "buddhist"
  | "washi"
  | "stationery"
  | "stonework"
  | "preciousStone"
  | "dolls"
  | "otherCrafts"
  | "craftMaterials"

export interface QuizQuestion {
  id: number
  craft: CraftCategory
  category: KnowledgeCategory
  format: QuestionFormat
  difficulty: Difficulty
  question: string
  options?: string[]
  answer: string
  explanation: string
  imageUrl?: string
}